+++ src/PhpSpreadsheet/Worksheet/AutoFilter.php
@@ -89,11 +89,8 @@ class AutoFilter
      */
     public function setRange($pRange)
     {
-        // Uppercase coordinate
-        $cellAddress = explode('!', strtoupper($pRange));
-        if (count($cellAddress) > 1) {
-            list($worksheet, $pRange) = $cellAddress;
-        }
+        // extract coordinate
+        list($worksheet, $pRange) = Worksheet::extractSheetTitle($pRange, true);
 
         if (strpos($pRange, ':') !== false) {
             $this->range = $pRange;