+++ src/PhpSpreadsheet/Writer/Xls/Parser.php
@@ -3,6 +3,7 @@
 namespace PhpOffice\PhpSpreadsheet\Writer\Xls;
 
 use PhpOffice\PhpSpreadsheet\Shared\StringHelper;
+use PhpOffice\PhpSpreadsheet\Worksheet\Worksheet as PhpspreadsheetWorksheet;
 use PhpOffice\PhpSpreadsheet\Writer\Exception as WriterException;
 
 // Original file header of PEAR::Spreadsheet_Excel_Writer_Parser (used as the base for this class):
@@ -643,7 +644,7 @@ class Parser
     private function convertRange3d($token)
     {
         // Split the ref at the ! symbol
-        list($ext_ref, $range) = explode('!', $token);
+        list($ext_ref, $range) = PhpspreadsheetWorksheet::extractSheetTitle($token, true);
 
         // Convert the external reference part (different for BIFF8)
         $ext_ref = $this->getRefIndex($ext_ref);
@@ -695,7 +696,7 @@ class Parser
     private function convertRef3d($cell)
     {
         // Split the ref at the ! symbol
-        list($ext_ref, $cell) = explode('!', $cell);
+        list($ext_ref, $cell) = PhpspreadsheetWorksheet::extractSheetTitle($cell, true);
 
         // Convert the external reference part (different for BIFF8)
         $ext_ref = $this->getRefIndex($ext_ref);